[refactor] #7 UserRepository 패키지 위치 변경
