+++ app/src/main/java/com/yhj/gdscandroidstudy/ui/edit/EditViewModel.kt
@@ -2,7 +2,7 @@ package com.yhj.gdscandroidstudy.ui.edit
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.yhj.gdscandroidstudy.data.user.UserRepository
+import com.yhj.gdscandroidstudy.domain.UserRepository
 import com.yhj.gdscandroidstudy.util.SUBSCRIPTION_TIMEOUT
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
